xtts_v2/pytorch/requirements.txt
coqui-tts
torchaudio
